#21 회원가입 로직 뷰모델로 이동
